Refining things on Windows, again.
- Added the GBECommon library
- Renamed Example3-Uniforms/Shaders to Example3-Uniforms/Resources
- Fixed up runtime paths and working directories on the projects
- Quieted some warnings

+++ GBECommon/Source/GBE_3DMath.c
@@ -184,7 +184,7 @@ GBE_Matrix4x4 GBE_Matrix4x4RotateAxisAngle(GBE_Vector3 axis, float angleInRadian
 
 GBE_Matrix4x4 GBE_Matrix4x4Perspective(float aspect, float fovy, float near, float far)
 {
-    float yScale = 1 / tan(fovy * 0.5f);
+    float yScale = 1 / tanf(fovy * 0.5f);
     float xScale = yScale / aspect;
     float zRange = far - near;
     float zScale = -(far + near) / zRange;
